estudos 2023

+++ estudosflutter2/lib/screens/home_screen/widgets/journal_card.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_webapi_first_course/helpers/weekday.dart';
 import 'package:flutter_webapi_first_course/models/journal.dart';
+import 'package:flutter_webapi_first_course/screens/home_screen/commom/confirmation_dialog.dart';
 import 'package:flutter_webapi_first_course/services/journal_service.dart';
 import 'package:uuid/uuid.dart';
 
@@ -143,20 +144,34 @@ class JournalCard extends StatelessWidget {
   }
 
   removeJournal(BuildContext context) {
-    JournalService service = JournalService();
+    showConfirmationDialog(
+      context,
+      content:
+          "Deseja realmente deletar essa tarefa do dia ${WeekDay(journal!.createdAt)}?",
+      affirmativeOption: "Remover!",
+    ).then((value) {
+      if(value != null)
+      {
+        if(value)
+        {
+          JournalService service = JournalService();
 
-    if (journal != null) {
-      service.remove(journal!.id).then((value) {
-        if (value) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text("Removido com sucesso"),
-            ),
-          );
+          if (journal != null) {
+            service.remove(journal!.id).then((value) {
+              if (value) {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  const SnackBar(
+                    content: Text("Removido com sucesso"),
+                  ),
+                );
 
-          refreshFunction();
+                refreshFunction();
+              }
+            });
+          }
         }
-      });
-    }
+      }
+    });
+
   }
 }
